+++ packages/impulse/lib/impulse.dart
@@ -6,5 +6,4 @@ export 'src/box.dart';
 export 'src/interfaces.dart';
 export 'src/reactivity_delegate.dart';
 export 'src/reference.dart';
-export 'src/scope.dart';
-export 'src/error_handling.dart';
+export 'src/async_utils.dart';