+++ packages/impulse/test/scope_test.dart
@@ -5,18 +5,24 @@ import 'package:impulse/impulse.dart';
 final dummyRef = Ref<Object>((store) => Object());
 
 void main() {
-  test('Ref is removed after scope', () async {
+  test('Ref is removed after async scope', () async {
     var store = Store();
 
-    await withScope(
-      (store) async {
-        await Future.microtask(() {});
-        store.init(dummyRef());
-      },
-      store: store,
-      ref: dummyRef(),
-    );
+    await store.withScope(dummyRef, (store, value) async {
+      await Future.microtask(() {});
+      store.init(dummyRef);
+    });
 
-    expect(store.exists(dummyRef()), isFalse);
+    expect(store.exists(dummyRef), isFalse);
+  });
+
+  test('Ref is removed after sync scope', () async {
+    var store = Store();
+
+    await store.withScope(dummyRef, (store, value) {
+      store.init(dummyRef);
+    });
+
+    expect(store.exists(dummyRef), isFalse);
   });
 }