+++ packages/impulse_flutter/lib/impulse_flutter.dart
@@ -5,4 +5,3 @@ export 'package:impulse/impulse.dart' hide $store;
 export 'src/flutter_delegate.dart';
 export 'src/store_scope.dart';
 export 'src/context_extensions.dart';
-export 'src/ref_extension.dart';